[2026-01-31] #7 HTTP 통신과 비동기 코드
